# Use python-isal's igzip_threaded module to replace igzip piping.