--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -47,6 +47,10 @@ jobs:
           optional-deps: false
         - os: ubuntu-20.04
           python-version: "3.10"
+          with-python-isal: false
+          optional-deps: true
+        - os: ubuntu-20.04
+          python-version: "3.10"
           optional-deps: false
           with-python-isal: false
           with-zstandard: true