--- a/setup.cfg
+++ b/setup.cfg
@@ -13,13 +13,12 @@ classifiers =
     Programming Language :: Python :: 3
 
 [options]
-python_requires = >=3.7
+python_requires = >=3.8
 package_dir =
     =src
 packages = find:
 install_requires =
     isal>=1.4.1; platform.machine == "x86_64" or platform.machine == "AMD64" or platform.machine == "aarch64"
-    typing_extensions; python_version<'3.8'
 
 [options.packages.find]
 where = src