--- a/src/xopen/__init__.py
+++ b/src/xopen/__init__.py
@@ -1053,9 +1053,10 @@ def _open_gz(filename, mode: str, compresslevel, threads, **text_mode_kwargs):
             return igzip_threaded.open(
                 filename,
                 mode,
-                compresslevel or isal_zlib.ISAL_DEFAULT_COMPRESSION,
+                isal_zlib.ISAL_DEFAULT_COMPRESSION if compresslevel is None
+                else compresslevel,
                 **text_mode_kwargs,
-                threads=threads or 1,
+                threads=1 if threads is None else threads,
             )
         except ValueError:  # Wrong compression level
             pass