--- a/src/xopen/__init__.py
+++ b/src/xopen/__init__.py
@@ -1056,7 +1056,7 @@ def _open_gz(filename, mode: str, compresslevel, threads, **text_mode_kwargs):
     assert mode in ("rt", "rb", "wt", "wb", "at", "ab")
     # With threads == 0 igzip_threaded defers to igzip.open, but that is not
     # desirable as a reproducible header is required.
-    if threads != 0:
+    if igzip_threaded and threads != 0:
         try:
             return igzip_threaded.open(  # type: ignore
                 filename,
@@ -1067,9 +1067,9 @@ def _open_gz(filename, mode: str, compresslevel, threads, **text_mode_kwargs):
                 **text_mode_kwargs,
                 threads=1 if threads is None else threads,
             )
-        except (AttributeError, ValueError):
-            # Igzip_threaded is None or wrong compression level
+        except ValueError:  # Wrong compression level
             pass
+    if threads != 0:
         try:
             if "r" in mode:
                 return _open_external_gzip_reader(