--- a/tox.ini
+++ b/tox.ini
@@ -21,6 +21,11 @@ deps =
     {[testenv]deps}
     zstandard
 
+[testenv:no-isal]
+commands=
+    pip uninstall -y isal
+    {[testenv]commands}
+
 [testenv:black]
 basepython = python3.10
 deps = black==22.3.0