@rhpvorderman 2023-10-12
I had to make a change requiring at least 3.8. This is due to python-isal having the same requirement. Given that 3.7 is out of support currently I think this is a valid change. Generally I'd like to maintain support as long as it is not annoying to do so. Given the extra typing requirements for 3.7 I think it is convenient to get rid of it.