@marcelm 2023-10-18
I looked into why coverage goes down. We now [never call _open_external_gzip_reader](https://app.codecov.io/gh/pycompression/xopen/pull/131/blob/src/xopen/__init__.py#L1027) because the isal import never fails. We have an extra test in the test matrix with `optional-dependencies: false` where we don’t install the system isal libraries, but I guess that does not have an effect because `python-isal` comes with libisal already, is that correct?

I’m ok with leaving it as-is (with decreased coverage) since the platforms we care about are covered. We can consider re-adding tests later (or possibly decide that we want isal to be a hard dependency?).

Can you please:

- [ ] Check that the description in the README regarding what happens when is still correct with this PR?
- [ ] Add a changelog entry.

Then feel free to merge and make a release. I guess this should be called 1.8.0, but I’ll leave that up to you.

Thanks!