@marcelm 2023-10-18
Looks good, thanks for fixing the coverage issue. Agreed that we shoul consider reducing the number of external readers.